move cacheable request to Traits namespace
