+++ tests/unit/API/CacheableRequestTraitTest.php
@@ -2,7 +2,7 @@
 
 namespace API;
 
-use SkyVerge\WooCommerce\PluginFramework\v5_10_10\API\Cacheable_Request_Trait;
+use SkyVerge\WooCommerce\PluginFramework\v5_10_10\API\Traits\Cacheable_Request_Trait;
 
 if ( ! defined( 'ABSPATH' ) ) {
 	define( 'ABSPATH', true );
@@ -20,7 +20,7 @@ class CacheableRequestTraitTest extends \Codeception\Test\Unit
 
 		require_once('woocommerce/api/interface-sv-wc-api-request.php');
 		require_once('woocommerce/api/abstract-sv-wc-api-json-request.php');
-		require_once('woocommerce/api/Cacheable_Request_Trait.php');
+		require_once('woocommerce/api/traits/Cacheable_Request_Trait.php');
 	}
 
 	/** @see CacheableRequest::get_cache_lifetime() */
